separated assessment related views
